requirements.txt
# Core Dependencies
mistralai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
httpx>=0.24.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# UI Options
streamlit>=1.28.0
gradio>=4.0.0

# Database & Caching
redis>=5.0.0
sqlalchemy>=2.0.0

# Utilities
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
numpy>=1.24.0
scikit-learn>=1.3.0
duckduckgo-search>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
